Fix analytics 500 from decimal duration values
A manual SQL UPDATE to cancel stuck executions stored duration
as decimal strings (via EXTRACT(EPOCH) * 1000). The analytics
and metrics queries used CAST(duration AS INTEGER), which fails
on decimal text like "1234567.890" in Postgres.

- Change all CAST(duration AS INTEGER) to CAST(duration AS
  NUMERIC) so decimal strings parse correctly
- Add migration 0026 to round existing decimal values back to
  integer strings

+++ keeperhub/lib/analytics/queries.ts
@@ -166,7 +166,7 @@ async function getWorkflowCounts(
       total: count(),
       success: sql<number>`SUM(CASE WHEN ${workflowExecutions.status} = 'success' THEN 1 ELSE 0 END)`,
       error: sql<number>`SUM(CASE WHEN ${workflowExecutions.status} IN ('error', 'cancelled') THEN 1 ELSE 0 END)`,
-      durationSum: sql<number>`COALESCE(SUM(CAST(${workflowExecutions.duration} AS INTEGER)), 0)`,
+      durationSum: sql<number>`COALESCE(SUM(CAST(${workflowExecutions.duration} AS NUMERIC))::INTEGER, 0)`,
       durationCount: sql<number>`SUM(CASE WHEN ${workflowExecutions.duration} IS NOT NULL THEN 1 ELSE 0 END)`,
     })
     .from(workflowExecutions)
